parameterizedFields folder: remove "parameterized fields" clause, rename file
