--- test/unit/operations/write/parameterizedFields/indirectEditEntityOfParameterizedField.ts
@@ -1,75 +1,0 @@
-import { CacheContext } from '../../../../../src/context';
-import { GraphSnapshot } from '../../../../../src/GraphSnapshot';
-import { nodeIdForParameterizedValue } from '../../../../../src/operations/SnapshotEditor';
-import { write } from '../../../../../src/operations/write';
-import { NodeId, StaticNodeId } from '../../../../../src/schema';
-import { query, strictConfig } from '../../../../helpers';
-
-const { QueryRoot: QueryRootId } = StaticNodeId;
-
-// These are really more like integration tests, given the underlying machinery.
-//
-// It just isn't very fruitful to unit test the individual steps of the write
-// workflow in isolation, given the contextual state that must be passed around.
-describe(`operations.write`, () => {
-
-  const context = new CacheContext(strictConfig);
-  const empty = new GraphSnapshot();
-  const viewerQuery = query(`{
-    viewer {
-      id
-      name
-    }
-  }`);
-
-  describe(`parameterized fields`, () => {
-
-    describe(`indirectly updating entity field with a direct reference`, () => {
-
-      let baseline: GraphSnapshot, snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>, parameterizedId: NodeId;
-      beforeAll(() => {
-        const parameterizedQuery = query(`query getAFoo($id: ID!) {
-          foo(id: $id, withExtra: true) {
-            id name extra
-          }
-        }`, { id: 1 });
-
-        parameterizedId = nodeIdForParameterizedValue(QueryRootId, ['foo'], { id: 1, withExtra: true });
-
-        const baselineResult = write(context, empty, parameterizedQuery, {
-          foo: {
-            id: 1,
-            name: 'Foo',
-            extra: false,
-          },
-        });
-        baseline = baselineResult.snapshot;
-
-        const result = write(context, baseline, viewerQuery, {
-          viewer: {
-            id: 1,
-            name: 'Foo Bar',
-          },
-        });
-        snapshot = result.snapshot;
-        editedNodeIds = result.editedNodeIds;
-      });
-
-      it(`updates the node for the field`, () => {
-        expect(snapshot.getNodeData(parameterizedId)).to.deep.eq({ id: 1, name: 'Foo Bar', extra: false });
-      });
-
-      it(`ensures normalized references`, () => {
-        const entity = snapshot.getNodeData('1');
-        expect(snapshot.getNodeData(QueryRootId).viewer).to.eq(entity);
-        expect(snapshot.getNodeData(parameterizedId)).to.eq(entity);
-      });
-
-      it(`marks only the entity as edited`, () => {
-        expect(Array.from(editedNodeIds)).to.have.members([QueryRootId, '1']);
-      });
-
-    });
-  });
-
-});
