+++ test/unit/operations/write/parameterizedFields/topLevelParameterizedReference.ts
@@ -2,7 +2,6 @@ import * as _ from 'lodash';
 
 import { CacheContext } from '../../../../../src/context';
 import { GraphSnapshot } from '../../../../../src/GraphSnapshot';
-import { ParameterizedValueSnapshot } from '../../../../../src/nodes';
 import { nodeIdForParameterizedValue } from '../../../../../src/operations/SnapshotEditor';
 import { write } from '../../../../../src/operations/write';
 import { NodeId, StaticNodeId } from '../../../../../src/schema';
@@ -19,20 +18,22 @@ describe(`operations.write`, () => {
   const context = new CacheContext(strictConfig);
   const empty = new GraphSnapshot();
 
-  describe(`top-level non-entity parameterized field`, () => {
+  describe(`top-level parameterized reference`, () => {
 
     let snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>, parameterizedId: NodeId;
     beforeAll(() => {
-      const parameterizedQuery = query(`query getAFoo($id: ID!) {
-        foo(id: $id, withExtra: true) {
-          name extra
-        }
-      }`, { id: 1 });
+      const parameterizedQuery = query(`
+        query getAFoo($id: ID!) {
+          foo(id: $id, withExtra: true) {
+            id name extra
+          }
+        }`, { id: 1 });
 
       parameterizedId = nodeIdForParameterizedValue(QueryRootId, ['foo'], { id: 1, withExtra: true });
 
       const result = write(context, empty, parameterizedQuery, {
         foo: {
+          id: 1,
           name: 'Foo',
           extra: false,
         },
@@ -41,8 +42,12 @@ describe(`operations.write`, () => {
       editedNodeIds = result.editedNodeIds;
     });
 
-    it(`writes a node for the field`, () => {
-      expect(snapshot.getNodeData(parameterizedId)).to.deep.eq({ name: 'Foo', extra: false });
+    it(`writes a node for the new entity`, () => {
+      expect(snapshot.getNodeData('1')).to.deep.eq({ id: 1, name: 'Foo', extra: false });
+    });
+
+    it(`writes a node for the field that points to the entity's value`, () => {
+      expect(snapshot.getNodeData(parameterizedId)).to.eq(snapshot.getNodeData('1'));
     });
 
     it(`creates an outgoing reference from the field's container`, () => {
@@ -55,16 +60,22 @@ describe(`operations.write`, () => {
       expect(values.inbound).to.deep.eq([{ id: QueryRootId, path: ['foo'] }]);
     });
 
+    it(`creates an outgoing reference from the parameterized field to the referenced entity`, () => {
+      const values = snapshot.getNodeSnapshot(parameterizedId)!;
+      expect(values.outbound).to.deep.eq([{ id: '1', path: [] }]);
+    });
+
+    it(`creates an incoming reference from the parameterized field to the referenced entity`, () => {
+      const entity = snapshot.getNodeSnapshot('1')!;
+      expect(entity.inbound).to.deep.eq([{ id: parameterizedId, path: [] }]);
+    });
+
     it(`does not expose the parameterized field directly from its container`, () => {
       expect(_.get(snapshot.getNodeData(QueryRootId), 'foo')).to.eq(undefined);
     });
 
-    it(`marks only the new field as edited`, () => {
-      expect(Array.from(editedNodeIds)).to.have.members([parameterizedId]);
-    });
-
-    it(`emits a ParameterizedValueSnapshot`, () => {
-      expect(snapshot.getNodeSnapshot(parameterizedId)).to.be.an.instanceOf(ParameterizedValueSnapshot);
+    it(`marks the new field and entity as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members([parameterizedId, '1']);
     });
 
   });
